Modificación de la clase pacientes, y migración en la base datos para eliminar la columna edad, debido a que esta solo será representada como un cálculo en el dto de respuesta, teniendo en cuenta la fecha de nacimiento. Evitando así guardar un valor estático en la base de datos, que sería edad, se utilizó la anotación @Transient en la variable edad para que no sea agregada como columna, y se modifica su get.

+++ JannieVet/src/main/java/com/brayandvlp/JannieVet/domain/mascotaPaciente/dtos/DatosRespuestaPaciente.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 public record DatosRespuestaPaciente(
         Long id,
         String nombre,
-        Integer edad,
+        Double edad,
         Especie especie,
         String raza,
         Double peso,
